3rd Place/requirements.txt
opencv-python
pandas
jupyter
tensorflow-gpu==2.1.0
pillow
matplotlib
lightgbm
sklearn